+++ cmd/rqlited/main.go
@@ -72,7 +72,6 @@ var noVerify bool
 var expvar bool
 var dsn string
 var inMem bool
-var disRedirect bool
 var showVersion bool
 var cpuprofile string
 
@@ -89,7 +88,6 @@ func init() {
 	flag.BoolVar(&expvar, "expvar", true, "Serve expvar data on HTTP server")
 	flag.StringVar(&dsn, "dsn", "", `SQLite DSN parameters. E.g. "cache=shared&mode=memory"`)
 	flag.BoolVar(&inMem, "mem", false, "Use an in-memory database")
-	flag.BoolVar(&disRedirect, "noredir", true, "Disable leader-redirect")
 	flag.BoolVar(&showVersion, "version", false, "Show version information and exit")
 	flag.StringVar(&cpuprofile, "cpuprofile", "", "Write CPU profile to file")
 	flag.Usage = func() {
@@ -216,7 +214,6 @@ func main() {
 
 	s.CertFile = x509Cert
 	s.KeyFile = x509Key
-	s.DisableRedirect = disRedirect
 	s.Expvar = expvar
 	s.BuildInfo = map[string]interface{}{
 		"commit":     commit,